CustomerControl Class
Completed the customer control class so the customers can complete some of the required functions of the system allowing the users to order, modify orders, cancel orders and more

+++ src/main/java/main/Meal.java
@@ -60,7 +60,7 @@ public class Meal {
         this.name = name;
     }
 
-    public double getPrice() {
+    public float getPrice() {
         return price;
     }
